include .text/.rela/func sym bundle if rela entry references changed local string
Consider a file the following single function:

void test_func() {
	printf("this is before");
}

with a patch that changes the contents of the string.

The result is that the .text, .rela, and function symbol information
for that function remain unchanged as the addend into the
.rodata.str1.1 section remains 0.  However the string that is
referenced by the rela entry has changed.

If the rela entry references a string section and the referenced
string has changed, the .text, .rela, and function symbol should be
marked as changed.  However, the current code does not do this.

This commit corrects this issue.

Signed-off-by: Seth Jennings <sjenning@redhat.com>