create-diff-object: generic special section support
Abstract out the common functionality for dealing with special sections
into a new kpatch_regenerate_special_sections() function.

The base sections are copied whole.  That's much easier than trying to
filter them, and any unused data is harmless.

The rela sections are partitioned into "groups" of relas.  Only those
groups which reference a changed function are kept.  The only difference
in the logic for handling each special section is determining the number
of rela entries in a given group.  Each section has its own group_size()
callback for this.

I used a callback instead of an int because some of the
soon-to-be-supported special sections require that their groups' sizes
be dynamically determined.