add support for manually ignore functions
This commit adds the __kpatch_ignore macro for ignore functions that
may change as a side effect of a change in another function.  The WARN
class of macros, for example, embed the line number in an instruction,
which will cause the function to be detected as changed when, in fact,
there has been no functional change.

Just at __kpatch_ignore to the function definition (like you would
__init) to cause the function to be ignored.

Signed-off-by: Seth Jennings <sjenning@redhat.com>